% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torchgnn-package.R
\docType{package}
\name{torchgnn-package}
\alias{torchgnn}
\alias{torchgnn-package}
\title{torchgnn: Graph Neural Networks for 'torch'}
\description{
Build and train graph neural networks on top of the 'torch' deep learning framework. Message-passing layers are provided for the Graph Convolutional Network (GCN) of Kipf and Welling (2017) \doi{10.48550/arXiv.1609.02907}, the GraphSAGE architecture of Hamilton, Ying and Leskovec (2017) \doi{10.48550/arXiv.1706.02216}, the Graph Attention Network (GAT) of Velickovic and others (2018) \doi{10.48550/arXiv.1710.10903}, the Graph Isomorphism Network (GIN) of Xu and others (2019) \doi{10.48550/arXiv.1810.00826}, and the regional GCN of Guo and others (2025) \doi{10.1080/24694452.2025.2558661}. These layers are composed into ready-made node-level and graph-level models. Supporting utilities construct sparse adjacency matrices from edge lists, normalize them, pool node embeddings into graph-level representations, and split graphs into training, validation and test sets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://josiahparry.github.io/torchgnn/}
  \item \url{https://github.com/JosiahParry/torchgnn}
  \item Report bugs at \url{https://github.com/JosiahParry/torchgnn/issues}
}

}
\author{
\strong{Maintainer}: Josiah Parry \email{josiah.parry@gmail.com} (\href{https://orcid.org/0000-0001-9910-865X}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Josiah Parry \email{josiah.parry@gmail.com} (\href{https://orcid.org/0000-0001-9910-865X}{ORCID}) [copyright holder]
}

}
\keyword{internal}
