% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-adj.R
\name{adj_from_edgelist}
\alias{adj_from_edgelist}
\title{Create Sparse Adjacency Matrix from Edge List}
\usage{
adj_from_edgelist(from, to, weight = NULL, n = NULL, symmetric = TRUE)
}
\arguments{
\item{from}{Integer or character vector of source nodes}

\item{to}{Integer or character vector of target nodes}

\item{weight}{Numeric vector of edge weights}

\item{n}{Number of nodes}

\item{symmetric}{Make adjacency symmetric}
}
\value{
Sparse COO tensor with attributes \code{node_ids} and \code{id_map}
}
\description{
Create Sparse Adjacency Matrix from Edge List
}
