% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-gcn-degree-normalize.R
\name{gcn_normalize}
\alias{gcn_normalize}
\alias{adj_row_normalize}
\alias{add_graph_self_loops}
\title{Add self-loops to a graph}
\usage{
gcn_normalize(adj)

adj_row_normalize(adj)

add_graph_self_loops(adj)
}
\arguments{
\item{adj}{a sparse COO tensor of the adjacency matrix. Can be weighted.}
}
\description{
Add self-loops to a graph
}
