% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-gcn-general.R
\name{model_gcn_general}
\alias{model_gcn_general}
\title{Multi-layer Generalized GCN Model (Hamilton 2020)}
\usage{
model_gcn_general(
  in_features,
  hidden_dims,
  out_features,
  activation = nnf_relu,
  out_activation = NULL,
  dropout = 0,
  normalize = TRUE
)
}
\description{
Stacks multiple generalized GCN layers with neighbor/self weight separation.
Optional row-normalization is applied internally if \code{normalize = TRUE}.
}
