% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-adj.R
\name{nodes_to_tensor}
\alias{nodes_to_tensor}
\title{Convert Node Features to Tensor}
\usage{
nodes_to_tensor(nodes, adj = NULL, node_id = NULL)
}
\arguments{
\item{nodes}{Dataframe with node features}

\item{adj}{Adjacency matrix with \code{id_map} attribute}

\item{node_id}{Column name for node IDs}
}
\value{
Dense tensor of shape \verb{[n_nodes, n_features]}
}
\description{
Convert Node Features to Tensor
}
