% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torchgeo-package.R
\docType{package}
\name{torchgnn-package}
\alias{torchgnn}
\alias{torchgnn-package}
\title{torchgnn: Graph Neural Network Extensions for `torch`}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://josiahparry.github.io/torchgnn/}
}

}
\author{
\strong{Maintainer}: Josiah Parry \email{josiah.parry@gmail.com} (\href{https://orcid.org/0000-0001-9910-865X}{ORCID})

}
\keyword{internal}
